db:
Exposure `occupants` is now a float (since it can be a fractional value
due to time-of-day ratios and/or dwelling fractions.


Former-commit-id: 5d12d0e3a809318c1bb1f6703e40f2a48e857e62 [formerly bf42aff4f2b9697f72acfd3ef8007419dff2513d] [formerly bf42aff4f2b9697f72acfd3ef8007419dff2513d [formerly 592ac63e6aa472885d792793d9eb017fc24a94d7]]
Former-commit-id: 604c84c0213e2a67b46edc9967f6eb28cbb05cc6
Former-commit-id: 2f15a230cb333da0cdf8cfc6245319fcd8ff6020

+++ openquake/engine/db/models.py
@@ -2798,7 +2798,7 @@ class Occupancy(djm.Model):
 
     exposure_data = djm.ForeignKey("ExposureData")
     period = djm.TextField()
-    occupants = djm.IntegerField()
+    occupants = djm.FloatField()
 
     class Meta:
         db_table = 'riski\".\"occupancy'
